Event Grid New Functions
The Event Grid is now a list of resources to be added to the schedule
grid. Clicking New Event creates a new box in the list. Right Click
Context menu allows for deleting of boxes. Event Grid Auto slides the
boxes to maintain the list. New events added to end